Add files via upload
A simple python program where I test out using scipy and matplotlib to visualize the probability interval of a coin flipped a random amount of times and presents the info on a graph. Still working on visualizing the results in a more efficient manner and possibly using Pandas or seaborn.